+++ src/SbmFizikusToMqtt.SbmConnector/Configurations/SbmConfiguration.cs
@@ -9,4 +9,6 @@ public sealed record SbmConfiguration
     [Required] public required string Password { get; init; }
 
     [Required][Url] public required string BaseUrl { get; init; }
+
+    [Required] public required bool OutdoorWeatherEnabled { get; init; }
 }